Fixed minor issue with Byte and Short.

+++ src/main/scala/ch/ethz/acl/passera/unsigned/UByte.scala
@@ -6,5 +6,5 @@ case class UByte(override val byteValue: Byte) extends AnyVal with Serializable 
 
 object UByte {
   def MinValue = UByte(0)
-  def MaxValue = UByte(~0.toByte)
+  def MaxValue = UByte(~0)
 }
